testType: rename iperf/netperf modules for TestType
Rename, to make it clearer that these are merely the implementations of
a TestType. Most other parts of the code don't need to know the exact
types, and only deal wit the TestTypeHandler base class.
