+++ testTypeNetPerf.py
